+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.1.1]
+
+- Added update of settings when editing `go.mod` and `go.work`
+- Added entry to the configuration not only `GOPATH` and `GOROOT` , but also `GOBIN`
+- Added more description for the plugin
+
 ## [0.1.0]
 
 - Added basic ability to evaluate GOPATH and GOROOT env variables to support the Go plugin